Compilation for MSVC2013, thanks @dalkon
OTclient now compiles in "Microsoft Visual Studio 2013 Express for Windows Desktop"
All the needed libraries you can download at https://www.dropbox.com/s/2yfb1c763io8efy/otclient-msvc13-libs.zip
NOTE: You have to change VC++ Directories to the properly directories
NOTE: Latested MSVC 2013 or greated is required

+++ src/framework/stdext/cast.h
@@ -121,7 +121,7 @@ template<>
 inline bool cast(const std::string& in, float& f) {
     double d;
     if(cast(in, d)) {
-        f=d;
+        f=(float)d;
         return true;
     }
     return false;
