+++ src/framework/stdext/time.h
@@ -36,7 +36,7 @@ void microsleep(size_t us);
 struct timer {
 public:
     timer() { restart(); }
-    float elapsed_seconds() { return (stdext::micros() - m_start)/1000000.0; }
+    float elapsed_seconds() { return (float)((stdext::micros() - m_start)/1000000.0); }
     ticks_t elapsed_millis() { return (stdext::micros() - m_start)/1000; }
     ticks_t elapsed_micros() { return stdext::micros() - m_start; }
     void restart() { m_start = stdext::micros(); }
